quinnapp-voice-service/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# Pipecat framework
pipecat-ai==0.0.48

# AI services
groq==0.11.0
cartesia==1.1.0

# Database
supabase==2.9.0

# Environment
python-dotenv==1.0.1

# Audio processing
soundfile==0.12.1
numpy==2.1.3
